# Use batch-affine addition during buckets aggregation