--- a/bellman/src/batch_addition.rs
+++ b/bellman/src/batch_addition.rs
@@ -2,31 +2,50 @@ use std::collections::HashSet;
 
 use pairing::{CurveAffine, CurveProjective,};
 
+const BATCH_SIZE: usize = 256;
+
 pub struct BucketAdder<G: CurveAffine> {
     buckets: Vec<G>,
-    a: Vec<G>,
-    b: Vec<G>,
-    indexes: Vec<usize>,
-    count: usize,
-    busy: HashSet<usize>,
+    
+    batch_lhs: Vec<G>,
+    batch_rhs: Vec<G>,
+    batch_indexes: Vec<usize>,
+    batch_count: usize,
+    buckets_in_batch: HashSet<usize>,
     collisions: Vec<(G, usize)>,
-    buffer_size: usize,
+    batch_size: usize,
 }
 
 impl<G: CurveAffine> BucketAdder<G> {
-    pub fn new(c: u32, chunk: usize) -> BucketAdder<G> {
+    pub fn new(c: u32, _chunk: usize) -> BucketAdder<G> {
         let zero = G::zero();
-        let buffer_size = 256;
+        let batch_size = BATCH_SIZE;
         BucketAdder { 
             buckets: vec![zero; (1 << c) - 1], 
-            a: vec![zero; buffer_size], 
-            b: vec![zero; buffer_size], 
-            indexes: vec![0; buffer_size],
-            count: 0,
-            busy: HashSet::new(),
+            batch_lhs: vec![zero; batch_size], 
+            batch_rhs: vec![zero; batch_size], 
+            batch_indexes: vec![0; batch_size],
+            batch_count: 0,
+            buckets_in_batch: HashSet::new(),
             collisions: Vec::new(),
-            buffer_size,
+            batch_size,
         }
+    }
+
+    pub fn add_to_bucket(&mut self, base: G, bucket: usize) {
+        self.add(base, bucket);
+
+        if self.batch_count == self.batch_size || self.collisions.len() >= self.batch_size / 2  {
+            self.process_batch();
+        }
+    }
+
+    pub fn finalize(mut self) -> Vec<G> {
+        while !self.collisions.is_empty() {
+            self.process_batch();
+        }
+        self.process_batch();
+        self.buckets
     }
 
     fn add(&mut self, base: G, bucket: usize) {
@@ -39,7 +58,7 @@ impl<G: CurveAffine> BucketAdder<G> {
         }
         
         // bucket is already in buffer, postpone that addition
-        if self.busy.contains(&bucket) {
+        if self.buckets_in_batch.contains(&bucket) {
             self.collisions.push((base, bucket));
             return;
         }
@@ -54,38 +73,22 @@ impl<G: CurveAffine> BucketAdder<G> {
             return;
         }
 
-        self.a[self.count] = self.buckets[bucket];
-        self.b[self.count] = base;
-        self.indexes[self.count] = bucket;
-        self.busy.insert(bucket);
-        self.count += 1;
-    }
-
-    pub fn add_to_bucket(&mut self, base: G, bucket: usize) {
-        self.add(base, bucket);
-
-        if self.count == self.buffer_size || self.collisions.len() >= self.buffer_size / 2  {
-            self.process_buffer();
-        }
-    }
-
-    pub fn finalize(mut self) -> Vec<G> {
-        while !self.collisions.is_empty() {
-            self.process_buffer();
-        }
-        self.process_buffer();
-        self.buckets
+        self.batch_lhs[self.batch_count] = self.buckets[bucket];
+        self.batch_rhs[self.batch_count] = base;
+        self.batch_indexes[self.batch_count] = bucket;
+        self.buckets_in_batch.insert(bucket);
+        self.batch_count += 1;
     }
     
-    fn process_buffer(&mut self) {
-        CurveAffine::batch_addition_assign(&mut self.a[0..self.count], &self.b[0..self.count]);
+    fn process_batch(&mut self) {
+        CurveAffine::batch_addition_assign(&mut self.batch_lhs[0..self.batch_count], &self.batch_rhs[0..self.batch_count]);
 
-        for i in 0..self.count {
-            self.buckets[self.indexes[i]] = self.a[i];
+        for i in 0..self.batch_count {
+            self.buckets[self.batch_indexes[i]] = self.batch_lhs[i];
         }
 
-        self.count = 0;
-        self.busy.clear();
+        self.batch_count = 0;
+        self.buckets_in_batch.clear();
 
         let collisions = self.collisions.clone();
         self.collisions.clear();