--- a/bellman/src/lib.rs
+++ b/bellman/src/lib.rs
@@ -22,6 +22,7 @@ pub mod sonic;
 mod group;
 mod source;
 mod multiexp;
+mod batch_addition;
 
 #[cfg(test)]
 mod tests;