--- a/bellman/src/source.rs
+++ b/bellman/src/source.rs
@@ -32,7 +32,8 @@ pub trait Source<G: CurveAffine> {
     /// Skips `amt` elements from the source, avoiding deserialization.
     fn skip(&mut self, amt: usize) -> Result<(), SynthesisError>;
 
-    fn next(&mut self) -> Option<G>;
+    /// Returns the element from the source. Fails if the point is at infinity.
+    fn next(&mut self) -> Result<G, SynthesisError>;
 }
 
 impl<G: CurveAffine> SourceBuilder<G> for (Arc<Vec<G>>, usize) {
@@ -70,10 +71,18 @@ impl<G: CurveAffine> Source<G> for (Arc<Vec<G>>, usize) {
         Ok(())
     }
 
-    fn next(&mut self) -> Option<G> {
-        let b = self.0.get(self.1)?;
+    fn next(&mut self) -> Result<G, SynthesisError> {
+        if self.0.len() <= self.1 {
+            return Err(io::Error::new(io::ErrorKind::UnexpectedEof, "expected more bases when adding from source").into());
+        }
+
+        if self.0[self.1].is_zero() {
+            return Err(SynthesisError::UnexpectedIdentity)
+        }
+        
+        let p = self.0[self.1];
         self.1 += 1;
-        Some(*b)
+        Ok(p)
     }
 }