This PR is aiming at improving the MSM performance. It has been achieved by implementing batch-affine addition on the buckets aggregation step of the Pippenger Algorithm.

There are two main parts of this PR:
1. Implementation of batch-affine addition (CurveAffine::batch_add_assign).
_Description of batch-affine addition:_ https://allfi.gitbook.io/zk-notes/cryptography/primitives/elliptic-curves/batch-affine-addition
_Reference implementation:_ https://github.com/AztecProtocol/barretenberg/blob/3b365807c55c386fd39aa54f66f7d06bc8727993/cpp/src/barretenberg/ecc/groups/element_impl.hpp#L698
2. Implementation of BucketAdder.
_Description of Pippenger Algorithm_: https://allfi.gitbook.io/zk-notes/cryptography/primitives/elliptic-curves/pippenger-algorithm
_Idea of the implementation_: https://github.com/z-prize/2022-entries/tree/main/open-division/prize4-msm-wasm/yrrid

> We do not do any preprocessing or sorting of the scalars, instead, we use a simple bit array to determine if a bucket is already the target of a point addition. In the event of a collision, we add the point and bucket pair to a linked list for later processing. When we have processed 1024 unique buckets, or reached 128 collisions, we do the batch invert and second phase of the point adds. We repeat this process until all the scalars have been processed.

The code here looks a bit awkward but all my attempts to improve readability led to performance degradation. So if anyone has any suggestions I would be grateful.

**Results:**
* i7-10750H, BN256: MSM duration was reduced from ~2100 ns per multiplication to ~1450 ns per multiplication (~31 %)
* i7-10750H: Proving time in wasm was reduced from ~18.5 s to 15 s (~19%)
* iPhone 11: Proving time in wasm was reduced from ~51 s to 35.5 s (~30%)
* M1: Proving time in wasm was reduced from 8.3-9 s to 7.3-7.5 s (~12-17%)